[expo-brightness] Move Android code to expo-brightness
