+++ android/app/build.gradle
@@ -263,6 +263,7 @@ dependencies {
   api 'host.exp.exponent:expo-barcode-scanner-interface:2.0.0'
   api 'host.exp.exponent:expo-barcode-scanner:2.0.0'
   api 'host.exp.exponent:expo-blur:1.0.0'
+  api 'host.exp.exponent:expo-brightness:1.0.0'
   api 'host.exp.exponent:expo-camera-interface:2.0.0'
   api 'host.exp.exponent:expo-camera:2.0.0'
   api 'host.exp.exponent:expo-contacts:2.0.0'
