+++ android/app/src/main/java/host/exp/exponent/MainApplication.java
@@ -16,6 +16,7 @@ import expo.modules.av.AVPackage;
 import expo.modules.backgroundfetch.BackgroundFetchPackage;
 import expo.modules.barcodescanner.BarCodeScannerPackage;
 import expo.modules.blurview.BlurViewPackage;
+import expo.modules.brightness.BrightnessPackage;
 import expo.modules.camera.CameraPackage;
 import expo.modules.constants.ConstantsPackage;
 import expo.modules.contacts.ContactsPackage;
@@ -76,6 +77,7 @@ public class MainApplication extends ExpoApplication implements AppLoaderPackage
         new BackgroundFetchPackage(),
         new BarCodeScannerPackage(),
         new BlurViewPackage(),
+        new BrightnessPackage(),
         new CameraPackage(),
         new ConstantsPackage(),
         new ContactsPackage(),
