+++ android/expoview/build.gradle
@@ -204,6 +204,7 @@ dependencies {
   api project(':expo-barcode-scanner')
   api project(':expo-barcode-scanner-interface')
   api project(':expo-blur')
+  api project(':expo-brightness')
   api project(':expo-camera')
   api project(':expo-camera-interface')
   api project(':expo-contacts')
