+++ android/expoview/src/main/java/versioned/host/exp/exponent/ExperiencePackagePicker.java
@@ -15,6 +15,7 @@ import expo.modules.av.AVPackage;
 import expo.modules.backgroundfetch.BackgroundFetchPackage;
 import expo.modules.barcodescanner.BarCodeScannerPackage;
 import expo.modules.blurview.BlurViewPackage;
+import expo.modules.brightness.BrightnessPackage;
 import expo.modules.camera.CameraPackage;
 import expo.modules.constants.ConstantsPackage;
 import expo.modules.contacts.ContactsPackage;
@@ -54,6 +55,7 @@ public class ExperiencePackagePicker {
       new BackgroundFetchPackage(),
       new BarCodeScannerPackage(),
       new BlurViewPackage(),
+      new BrightnessPackage(),
       new CameraPackage(),
       new ConstantsPackage(),
       new ContactsPackage(),
