+++ packages/expo-brightness/android/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
-<manifest package="expo.modules.template">
+<manifest package="expo.modules.brightness">
 </manifest>
