+++ packages/expo-brightness/android/src/main/java/expo/modules/brightness/BrightnessPackage.java
@@ -7,16 +7,10 @@ import java.util.List;
 
 import expo.core.BasePackage;
 import expo.core.ExportedModule;
-import expo.core.ViewManager;
 
 public class BrightnessPackage extends BasePackage {
   @Override
   public List<ExportedModule> createExportedModules(Context context) {
     return Collections.singletonList((ExportedModule) new BrightnessModule(context));
-  }
-
-  @Override
-  public List<ViewManager> createViewManagers(Context context) {
-    return Collections.singletonList((ViewManager) new BrightnessViewManager());
   }
 }
