--- packages/expo-brightness/android/src/main/java/expo/modules/brightness/BrightnessView.java
@@ -1,12 +1,0 @@
-package expo.modules.brightness;
-
-import android.content.Context;
-import android.view.View;
-
-import expo.core.ModuleRegistry;
-
-public class BrightnessView extends View {
-  public BrightnessView(Context context, ModuleRegistry moduleRegistry) {
-    super(context);
-  }
-}
