--- packages/expo-brightness/android/src/main/java/expo/modules/brightness/BrightnessViewManager.java
@@ -1,41 +1,0 @@
-package expo.modules.brightness;
-
-import android.content.Context;
-
-import java.util.Arrays;
-import java.util.List;
-
-import expo.core.ModuleRegistry;
-import expo.core.ViewManager;
-import expo.core.interfaces.ModuleRegistryConsumer;
-
-public class BrightnessViewManager extends ViewManager<BrightnessView> implements ModuleRegistryConsumer {
-  private static final String TAG = "ExpoBrightnessView";
-
-  private ModuleRegistry mModuleRegistry;
-
-  @Override
-  public String getName() {
-    return TAG;
-  }
-
-  @Override
-  public BrightnessView createViewInstance(Context context) {
-    return new BrightnessView(context, mModuleRegistry);
-  }
-
-  @Override
-  public ViewManagerType getViewManagerType() {
-    return ViewManagerType.SIMPLE;
-  }
-
-  @Override
-  public List<String> getExportedEventNames() {
-    return Arrays.asList("onSomethingHappened");
-  }
-
-  @Override
-  public void setModuleRegistry(ModuleRegistry moduleRegistry) {
-    mModuleRegistry = moduleRegistry;
-  }
-}
